+++ examples/hkt.c
@@ -29,63 +29,55 @@
 // clang-format off
 #define DefTreeG(branch, type)                                                 \
     choice(                                                                    \
-        TreeG(branch, type),                                                   \
-        variantMany(Branch(branch, type),                                      \
+        P(TreeG, branch, type),                                                \
+        variantMany(P(Branch, branch, type),                                   \
             field(data, type)                                                  \
             field(branches,                                                    \
-                POICA_MONOMORPHISE(branch, TreeG(branch, type))                \
+                P(branch, P(TreeG, branch, type))                              \
             )                                                                  \
         )                                                                      \
-        variant(Leaf(branch, type), type))
-// clang-format on
+        variant(P(Leaf, branch, type), type))
 
-#define TreeG(branch, type)  POICA_MONOMORPHISE(TreeG, branch, type)
-#define Branch(branch, type) POICA_MONOMORPHISE(Branch, branch, type)
-#define Leaf(branch, type)   POICA_MONOMORPHISE(Leaf, branch, type)
-
-// clang-format off
 #define DefBinaryTree(type)                                                    \
     record(                                                                    \
-        BinaryTree(type),                                                      \
+        P(BinaryTree, type),                                                   \
         field(left, struct type *)                                             \
         field(right, struct type *)                                            \
     )
-// clang-format on
-#define BinaryTree(type) POICA_MONOMORPHISE(BinaryTree, type)
 
-// clang-format off
 #define DefWeirdTree(type)                                                     \
     record(                                                                    \
-        WeirdTree(type),                                                       \
+        P(WeirdTree, type),                                                    \
         field(text, const char *)                                              \
     )
 // clang-format on
-#define WeirdTree(type) POICA_MONOMORPHISE(WeirdTree, type)
 
-DefBinaryTree(TreeG(BinaryTree, int));
+DefBinaryTree(P(TreeG, BinaryTree, int));
 DefTreeG(BinaryTree, int);
 
-DefWeirdTree(TreeG(WeirdTree, int));
+DefWeirdTree(P(TreeG, WeirdTree, int));
 DefTreeG(WeirdTree, int);
 
 void binary_tree(void) {
-    TreeG(BinaryTree, int) _456_leaf = Leaf(BinaryTree, int)(456);
-    TreeG(BinaryTree, int) _789_leaf = Leaf(BinaryTree, int)(789);
+    P(TreeG, BinaryTree, int) _456_leaf = P(Leaf, BinaryTree, int)(456);
+    P(TreeG, BinaryTree, int) _789_leaf = P(Leaf, BinaryTree, int)(789);
 
-    TreeG(BinaryTree, int) binary_tree =
-        Branch(BinaryTree, int)(123,
-                                (BinaryTree(TreeG(BinaryTree, int))){
-                                    &_456_leaf,
-                                    &_789_leaf,
-                                });
+    P(TreeG, BinaryTree, int)
+    binary_tree =
+        P(Branch, BinaryTree, int)(123,
+                                   (P(BinaryTree, P(TreeG, BinaryTree, int))){
+                                       &_456_leaf,
+                                       &_789_leaf,
+                                   });
 }
 
 void weird_tree(void) {
-    TreeG(WeirdTree, int) weird_tree_1 =
-        Branch(WeirdTree, int)(123,
-                               (WeirdTree(TreeG(WeirdTree, int))){
-                                   .text = "Hey",
-                               });
+    P(TreeG, WeirdTree, int)
+    weird_tree_1 =
+        P(Branch, WeirdTree, int)(123,
+                                  (P(WeirdTree, P(TreeG, WeirdTree, int))){
+                                      .text = "Hey",
+                                  });
 }
 
 int main(void) {